Updated scripts, mostly tag updates
Immersive engineering updated its material tags to better comply with common form, using raw_materials instead of raw_ores, I updated my scripts to align with this.

Also the cutters list in the hammer recipes script wasn't a fan of tags, so just replaced it with individual items from the tags, as the tags were pretty small and only contained a few items in total.