Update BYG nether config
Adjusted the spawn heights of multiple biomes, so that way certain biomes only spawn on certain layers, with hotter biomes settle in the lowest layer, and the coldest/sparsest biomes appear on the upper layer, with the more temperate forests appearing in the middle.

For exact rates, see the file.