Added custom fluid mixins
Added a custom fluid mixin for every molten metal, that converts a source block into its respective metal block when it comes in contact with water.

Also added a rule for all flowing molten metals, that converts it into a custom slag block when it comes in contact with water.